sio1blade windows crash fix;
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@90125 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ plugins/bladesio1/connection.h
@@ -22,13 +22,13 @@
 
 /***************************************************************************/
 
-extern s32 connectionOpen();
+extern int connectionOpen();
 extern void connectionClose();
 
 /***************************************************************************/
 
-extern size_t connectionSend(u8 *pdata, s32 size);
-extern size_t connectionRecv(u8 *pdata, s32 size);
+extern int connectionSend(u8 *pdata, s32 size);
+extern int connectionRecv(u8 *pdata, s32 size);
 
 /***************************************************************************/
